requirements.txt
pbr>=0.6,!=0.7,<1.0
eventlet>=0.13.0
PyYAML>=3.1.0
pecan>=0.4.5
WSME>=0.6
amqplib>=0.6.1
argparse
croniter>=0.3.4 # MIT License
requests>=1.1
kombu>=2.4.8
oslo.config>=1.2.0
oslo.messaging>=1.3.0
paramiko>=1.9.0
python-keystoneclient>=0.7.0
networkx>=1.8
six>=1.6.0
SQLAlchemy>=0.7.8,<=0.9.99
python-ceilometerclient>=1.0.6
python-keystoneclient>=0.8.0
python-novaclient>=2.17.0
CherryPy>=3.2.2